Bueno, esto es actualizacion despues de cursar le semsestre entero sin hacer commits, asi que muchas cosas cambiaron. En general hay ams estructura en la investrigacion, la separacion por etapas tiene ese proporsito. Tambien hay ahora la capeta servicio social donde hay todo lo referente y las clases del semestre cursado.

+++ investigacion/sobre-la-construccion/programas/midi-csv/filtro-silencios.py
@@ -24,7 +24,7 @@ for _, ins in instrumentos: # iteramos sobre cada instrumento
                 print(f"\naquí va un silencio que duraría {silencio}")
                 # guardamos el silencio
                 cancion_cc.append({
-                        'instrumento': float(ins.loc[i,'instrumento']),
+                        'instrumento': int(ins.loc[i,'instrumento']),
                         'nota': -1,
                         'volumen': 0.0,
                         'dura': float(silencio),
